made to throw exception when no user has this article

+++ SportsHub/SportsHub.Application/Services/ArticleService.cs
@@ -105,7 +105,13 @@ namespace SportsHub.AppService.Services
                 throw new NotFoundException(string.Format(ExceptionMessages.NotFound, ExceptionMessages.Article));
             }
 
-            _unitOfWork.UserRepository.DeleteArticle(articleForDelete);
+            var userHasArticle = await _unitOfWork.UserRepository.DeleteArticle(articleForDelete);
+
+            if (userHasArticle is false)
+            {
+                throw new NotFoundException(string.Format(ExceptionMessages.NotFound, ExceptionMessages.Article));
+            }
+
 
             _unitOfWork.ArticleRepository.DeleteArticle(articleForDelete);
             await _unitOfWork.SaveChangesAsync();
